requirements.txt
boxsdk>=2.7.1
jira>=2.0.0
PyYAML>=5.3.1
requests>=2.23.0
